블록이 벽에 부딛히도록 수정

+++ OpenSourceConsoleGame/소스.cpp
@@ -16,10 +16,13 @@
 #define GBOARD_ORIGIN_Y 2
 #define GBOARD_ORIGIN_X 4
 
+int GameBoard[GBOARD_HEIGHT + 1][GBOARD_WIDTH + 2];
 int block_id;
 int block_spin;
 int curPosX, curPosY;
 int speed = 30;
+
+
 
 void SetCurrentCursorPos(int x, int y)
 {
@@ -80,8 +83,24 @@ void deleteBlock(int blockInfo[4][4])
 	}
 }
 
+bool IsCollision(int nextX, int nextY, int nextR)
+{
+	nextX = (nextX - GBOARD_ORIGIN_X) / 2;
+	nextY = (nextY - GBOARD_ORIGIN_Y);
+
+	for (int y = 0; y < 4; y++)
+		for (int x = 0; x < 4; x++)
+		{
+			if (GameBoard[nextY + y][nextX + x] != 0 && blockModel[block_id][nextR][y][x] != 0)
+				return true;
+		}
+	return false;
+}
+
 void ShiftLeft()
 {
+	if (IsCollision(curPosX - 2, curPosY, block_spin))
+		return;
 	deleteBlock(blockModel[block_id][block_spin]);
 	curPosX -= 2;
 	SetCurrentCursorPos(curPosX, curPosY);
@@ -90,6 +109,8 @@ void ShiftLeft()
 
 void ShiftRight()
 {
+	if (IsCollision(curPosX + 2, curPosY, block_spin))
+		return;
 	deleteBlock(blockModel[block_id][block_spin]);
 	curPosX += 2;
 	SetCurrentCursorPos(curPosX, curPosY);
@@ -114,6 +135,8 @@ void BlockUp()
 
 void RotateBlock()
 {
+	if (IsCollision(curPosX, curPosY, block_spin + 1))
+		return;
 	deleteBlock(blockModel[block_id][block_spin]);
 	block_spin = (block_spin + 1) % 4;
 	SetCurrentCursorPos(curPosX, curPosY);
@@ -175,23 +198,71 @@ void drawBoard()
 	}
 }
 
+void ReDraw()
+{
+	for (int y = 0; y < GBOARD_HEIGHT; y++)
+	{
+		for (int x = 1; x < GBOARD_WIDTH+1; x++)
+		{
+			SetCurrentCursorPos(GBOARD_ORIGIN_X + (x * 2), GBOARD_ORIGIN_Y + y);
+			if (GameBoard[y][x] == 0)
+			{
+				printf("  ");
+			}
+			else
+			{
+				printf("бс");
+			}
+		}
+	}
+}
+
+COORD GetArrPos()
+{
+	COORD cur = GetCurrentCursorPos();
+	COORD arr;
+	arr.X = (cur.X - GBOARD_ORIGIN_X) / 2;
+	arr.Y = cur.Y - GBOARD_ORIGIN_Y;
+
+	return arr;
+}
+
+
+
+void initGameBoard()
+{
+	for(int y=0;y<GBOARD_HEIGHT + 1; y++)
+		for (int x = 0; x < GBOARD_WIDTH + 2; x++)
+		{
+			if (x == 0 || x == GBOARD_WIDTH + 1 || y == GBOARD_HEIGHT)
+				GameBoard[y][x] = 1;
+			else
+				GameBoard[y][x] = 0;
+		}
+}
+
 int main()
 {
+	initGameBoard();
 	block_spin = 0;
 	srand(time(NULL));
 	block_id = rand() % 7;
-	curPosX = curPosY = 0;
+	curPosX = 12;
+	curPosY = 0;
 	system("mode con: cols=60 lines=30");
 
 
 	RemoveCursor();
 
 	drawBoard();
-
+	ReDraw();
 	while (1)
 	{
-		BlockDown();
-		ProcessKeyInput();
+		while (IsCollision(curPosX, curPosY + 1, block_spin) == false)
+		{
+			BlockDown();
+			ProcessKeyInput();
+		}
 	}
 
 	getchar();
